# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sessio API",
        "description": "API para gerenciamento de sessões, clientes e bloqueios de tempo.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/time-blocks": {
            "get": {
                "tags": [
                    "TimeBlocks"
                ],
                "summary": "Listar todos os blocos de tempo do usuário autenticado",
                "operationId": "069896b70026ccc2898feadcf31ab748",
                "responses": {
                    "200": {
                        "description": "Lista de blocos de tempo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TimeBlock"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "TimeBlocks"
                ],
                "summary": "Criar um novo bloco de tempo",
                "operationId": "258a08b2f3bcec99b68b62e051a6a233",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Almoço"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-01-01T12:00:00-03:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-01-01T13:00:00-03:00"
                                    },
                                    "color_hex": {
                                        "type": "string",
                                        "example": "#FF5733",
                                        "nullable": true
                                    },
                                    "emoji": {
                                        "type": "string",
                                        "maxLength": 5,
                                        "example": "🍽️",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bloco de tempo criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TimeBlock"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Acesso negado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/time-blocks/{id}": {
            "get": {
                "tags": [
                    "TimeBlocks"
                ],
                "summary": "Exibir um bloco de tempo específico",
                "operationId": "f3f25a9f9e3c1fb6f8a3c2e070cc355b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do bloco de tempo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bloco de tempo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TimeBlock"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Acesso negado"
                    },
                    "404": {
                        "description": "Bloco de tempo não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "TimeBlocks"
                ],
                "summary": "Atualizar um bloco de tempo existente",
                "operationId": "fa71f98625fff5863830663fd837f975",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do bloco de tempo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Almoço Atualizado"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-01-01T12:30:00-03:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-01-01T13:30:00-03:00"
                                    },
                                    "color_hex": {
                                        "type": "string",
                                        "example": "#FF5733",
                                        "nullable": true
                                    },
                                    "emoji": {
                                        "type": "string",
                                        "maxLength": 5,
                                        "example": "🍽️",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bloco de tempo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TimeBlock"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Acesso negado"
                    },
                    "404": {
                        "description": "Bloco de tempo não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "TimeBlocks"
                ],
                "summary": "Excluir um bloco de tempo",
                "operationId": "4533505ce7dcf052c16d1d1352690b64",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do bloco de tempo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Bloco de tempo excluído com sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Acesso negado"
                    },
                    "404": {
                        "description": "Bloco de tempo não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "TimeBlock": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Almoço"
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00-03:00"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T13:00:00-03:00"
                    },
                    "color_hex": {
                        "type": "string",
                        "maxLength": 7,
                        "example": "#FF5733",
                        "nullable": true
                    },
                    "emoji": {
                        "type": "string",
                        "maxLength": 5,
                        "example": "🍽️",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "item_type": {
                        "type": "string",
                        "example": "block"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}